backend/deliver-document/requirements.txt
reportlab==4.2.2
pillow>=9.0.0
chardet
psycopg2-binary
